+++ src/app/preferences/customization/page.tsx
@@ -1,0 +1,5 @@
+const Page = () => {
+  return <div>page</div>;
+};
+
+export default Page;
