+++ src/app/preferences/result/_components/DrinkList.tsx
