+++ src/app/preferences/result/page.tsx
