+++ src/hooks/result/useGuestDrinkRecommendations.ts
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 
-import { recommendDrinks } from '@/app/result/action';
 import { Tables } from '@/types/supabase';
+import { recommendDrinks } from '@/utils/preference/action';
 
 const useGuestDrinkRecommendations = () => {
   const [drinks, setDrinks] = useState<Tables<'reco_results'>[] | null>(null);
